Commit 1
Here is a basic project outline. There is a DatabaseManager class which will create the SparkSession as well as a function for any DataFrames. It also utilizes the DatabaseAnalysis class to map functions contained within that class into a scenario function which maps each DataAnalysis function to its own case which can then be referenced in Main. There is a MenuChoices object with functions that can be called within Main. The Main object utilizes a while loop to control the flow of the application with each case calling either the relevant menu or DataAnalysis function. It is definitely a recursive menu that can be optimized. However, my main focus right now is connecting to Azure within the DatabaseManager class. Tek created the LoginMenu to check for a username/password combination which needs to be incorporated into Main.