Add files via upload
Unfortunately due to cost restraints as well as issues with active directory, we were not able to fully integrate with the Azure cloud. Our intention with the project was to conduct everything in the cloud - data storage as well as performing analytics. However, as students none of us were able to front the costs to get onto Databricks for analytics out of our own personal funds, and Revature was not able to properly configure the active directory in time to provide us with a company-funded databricks. However, this guide will get you to the point where your cleaned data is hosted on an Azure MySQL cloud database, which future iterations of this project can use.