Fully Function Menu Commit
All menus are contained in a MenuChoices Object which has functions to call login_2 from the Login class, as well as functions for a main_menu, Q1, Q2, Q3, and Q4. The Database Analysis class handles all of the queries which are then mapped to a case within the DatabaseManager class. An instance of databaseManager is then used to run each query from the MenuChoices class. DatabaseManager was also updated with a Timer function that is also utilized after each query to pause the application until a user presses any key.